Adjusted Health Decrease
Previously, when the enemy comes into contact with the player, the player's health decreased way too much.

Now, it only decreases slightly.

I also added a comment to remind myself to implement the struggle system later on (I recalled us talking about having the player struggle with the enemy and having to press a button a number of times before escaping).