Random Path Traversal (Enemy)
The enemy now traverses a randomly generated path.

I also added visuals for the path the enemy follows.

I also updated the GridManagerScript. The visual for seeing the entire node system is now centered at the position of the terrain (ground).

This process might become more complicated with Ians implementation of the map generation.