Struggle System Implemented
Implemented the quick time event system.

Not fully finished, but mostly done. I just need to make it so that the moment the player fails to succeed in the quick time event, they lose a heart.

Also, I need to make it so that the player faces the enemy and the enemy actually falls to the ground and not levitate in midair.