Simplified Decision Tree and Advanced Raycast FOV
Simplified the decision tree. I don't see any point in having an intermediate step/condition for now. It's always possible to easily expand it if we want to in the future.

I also made the raycasting system for advanced. Now it implements an actual vision system (not just shooting a raycast in front of the enemy).