Disable Item On Pickup
Instead of destroying an item on pick up, I just disable it. The reason for this approach is that, if I were to destroy the game object, I wouldn't be able to use it again (Unity frees up memory space when game object is destroyed, effectively removing reference of the game object).