Walkman Rotation, Item Slot Number, Disable Controls on Pause Menu
Rotated Walkman so when we add it to the inventory, it faces the camera.

Also, I added numbers for the item slots to let players know which buttons to click and trying to select an item in the item slots.

Finally, I moved the code for closing a note and for displaying instructions for picking up an item into the "Move" method. The "move" method is now only called when the timescale is greater than 0 (in other words, when we are not in the pause menu).