Adjusted Icon Image & Fixed Movement in Struggle
Adjusted the icon image in the inventory when an item is collected/picked up.

I also remembered why it was I checked for "player.enabled". StruggleSystem disables the player script, which means we should prevent the player from moving (i.e. check to see if "player.enabled" is true or false).