5 Second Chase Sequence
Enemy will chase the player for 5 seconds, even if the player is not direct in front of the enemy. We can, of course, increase this duration by changing the field "MAX_PURSUIT_TIME".

NOTE: This is currently being tested, if there are any bugs or weird issues, let me know and I'll work on fixing it.

The code itself is basically almost identical to the code for "IdleNode". Will test it more to see if we should even keep the 2 codes separate or if we should merge them into a single static class.