Drop & View Implemented
Items can now be picked up and dropped down.

Items can also be "viewed" or "Inspected" (currently called "view", but changing the text is easy).

Inspecting mode allows players to zoom in, zoom out, and rotate items.

There is a small issue/bug, when viewing an item, if the enemy tries to start a struggle mode, the struggle mode is not activated (not entirely sure why). Will look into it.