Activate Item, conversation, etc
Implemented the ability to "activate" an item. We would need to create separate scripts that inherit from "Activation" script for newer items that can activate in order for the functionality to work (dynamically).

I also implemented/pushed my conversation scripts. These scripts will allow us to display conversations between individuals (or a monologue if we want).

Finally, I forgot to push a change for the enemy 5 second movement. Now it should work. There is a small bug, which I will look into (sometimes the enemy tries to go to the path the player is at, but ignores the fact that the player is literally right in front of them).