Fixed Enemy Bug & Added Close Instruction
Turns out the issue was that the player collider was too small to detect sometimes, so I had to just simply adjust the radius (will test more just to be sure).

I also added a message at the very beginning of a "conversation" event that tells the player to click on "E" to close the "conversation" at anytime.